+++ pyArango/theExceptions.py
@@ -73,7 +73,7 @@ class InvalidDocument(pyArangoException) :
         pyArangoException.__init__(self, message, errors)
 
     def __str__(self) :
-        self.strErrors = []
+        strErrors = []
         for k, v in self.errors.items() :
             strErrors.append("%s -> %s" % (k, v))
         strErrors = '\n\t'.join(strErrors)
@@ -98,7 +98,12 @@ class AQLQueryError(pyArangoException) :
     def __init__(self, message, query, errors = None) :
         if errors is None :
             errors = {}
-        message = "Error in: %s.\n->%s" % (query, message)
+        lq = []
+        for i, ll in enumerate(query.split("\n")) :
+            lq.append("%s: %s" % (i+1, ll))
+        lq = '\n'.join(lq)
+
+        message = "Error in:\n%s.\n->%s" % (lq, message)
         pyArangoException.__init__(self, message, errors)
 
 class CursorError(pyArangoException):